requirements.txt
requests>=2.25.0
pywebview>=4.0.0
